+++ lib/cmdx/task.rb
@@ -214,8 +214,9 @@ module CMDx
       #
       # @param keys [Array<Symbol>]
       # @param options [Hash{Symbol => Object}] see {Output#initialize}
+      # @yield nested-output DSL block (see {Outputs::ChildBuilder})
       # @return [Outputs]
-      def outputs(*keys, **options)
+      def outputs(*keys, **options, &)
         @outputs ||=
           if superclass.respond_to?(:outputs)
             superclass.outputs.dup
@@ -225,7 +226,7 @@ module CMDx
 
         return @outputs if keys.empty?
 
-        @outputs.register(*keys, **options)
+        @outputs.register(*keys, **options, &)
       end
       alias output outputs